[refactor] 레스토랑 리스트 리팩토링
- 레스토랑 정보, 아이콘, 리스트아이템 컴포넌트 분리
- 컴포넌트 이름 변경
- 레스토랑 리스트 필터 로직을 훅으로 분리
